Removes building an intermediate service provider at startup, which is considered an anti-pattern, leading to occasional compatibility issues. This unblocks use in [Aspire](https://learn.microsoft.com/en-us/dotnet/aspire/) (more specifically, its usage of `AddNpgsqlDataSource`, which throws an `ObjectDisposedException`) and fixes #1082.
